*** empty log message ***

+++ kernel/linux/include/linux/netfilter_bridge/ebt_nat.h
@@ -4,8 +4,8 @@
 struct ebt_nat_info
 {
 	unsigned char mac[ETH_ALEN];
-	// EBT_ACCEPT, EBT_DROP, EBT_CONTINUE or EBT_RETURN
-	int target;
+	// EBT_ACCEPT, EBT_DROP or EBT_CONTINUE
+	__u8 target;
 };
 #define EBT_SNAT_TARGET "snat"
 #define EBT_DNAT_TARGET "dnat"
